[FE] Tidy up packages importing and add comments
