+++ frontend/src/Components/ui/Logo.js
@@ -1,3 +1,4 @@
+// React Utils, UI Components
 import { Box, Center } from "@chakra-ui/react";
 
 const Logo = (props) => {
