+++ frontend/src/Containers/Game/GuessTable.js
@@ -1,3 +1,4 @@
+// React Utils, UI Components
 import React, { useState, useEffect } from "react";
 import { useNavigate, useLocation } from "react-router-dom";
 import { Box, Stack, HStack, Text, Divider } from "@chakra-ui/react";
