+++ frontend/src/Containers/utils/redact.js
@@ -1,5 +1,7 @@
 import { commonWords, marks } from "./variables";
 
+// Function for redacting given content with guesses and commonWords revealed
+// Return the redacted content
 function redact(content, guesses) {
   let redacted = "";
   const words = content.split(" ");
